Use QImage to download images from the server
We now check what the content type of the request is and determine if
we're downloading some JSON query response or an image we requested.
Currently we just save the image out but later we'll want to grab the
blurred and unblurred images and composite them on a background thread.