Move background image construction into app code
The testing code in main has been migrated into a background task we run
in Qt that now operates on QImages. Qt receives the images from the
server as RGB32 so I've also switched to operate on a hard-coded little
endian RGB32 format. I'm not sure if anyone will use this on a big
endian machine, but they'll definitely notice something is wrong :).

If the endianness fix is straightforward I'll switch it. It might be as
simple as having the channel loop just go over all 4 channels instead of
the first 3.