Starting to play with UI elements some
Added button to force updating the background now. Although this is currently
just hardcoded to set it to some image. Also looking into list items so
we can have selections for which artists we want and so on. Should probably
look into background tasks now though and how to communicate with them
from the main app.